requirements.txt
asgiref==3.8.1
awsebcli==3.20.10
black==24.8.0
blessed==1.20.0
boto3==1.28.0
botocore==1.31.85
cement==2.8.2
certifi==2024.7.4
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.3
distlib==0.3.8
Django==5.1
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-filter==24.3
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.15.4
flake8==7.1.1
identify==2.6.0
idna==3.7
inflection==0.5.1
jmespath==1.0.1
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.10.1
pillow==10.4.0
platformdirs==4.2.2
psycopg2-binary==2.9.9
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.6.2
semantic-version==2.8.5
six==1.16.0
sqlparse==0.5.1
termcolor==1.1.0
tomli==2.0.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.19
virtualenv==20.26.3
wcwidth==0.1.9
